+++ examples/windows-recipe-writefile-advanced.rs
@@ -295,7 +295,7 @@ where
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let (session, profile) = common::create_vmi_session()?;
+    let (session, _profile) = common::create_vmi_session()?;
 
     let explorer_pid = {
         // This block is used to drop the pause guard after the PID is found.
@@ -331,7 +331,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     session.handle(|session| {
         InjectorHandler::new(
             session,
-            &profile,
             explorer_pid,
             recipe_factory(GuestFile::new(
                 "C:\\Users\\John\\Desktop\\test.txt",